+++ database/migrations/2022_05_21_095151_create_skills_table.php
@@ -13,15 +13,17 @@ class CreateSkillsTable extends Migration
      */
     public function up()
     {
-        Schema::create('skills', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('category_id')
-                ->constrained()
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table->string('title')->index();
-            $table->text('description')->index();
-        });
+        if (! Schema::hasTable('skills')) {
+            Schema::create('skills', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('category_id')
+                    ->constrained()
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table->string('title')->index();
+                $table->text('description')->index();
+            });
+        }
     }
 
     /**
